feat(build): keep buildcore.sh from moving binary packages in source-only mode
buildcore.sh calls makerpm and then moves the packages that makerpm made
into the delivery directories. A source-only build makes no binary package,
thus the move of the binary package finds nothing and reports an error.

The error is not the only result. Before each move, buildcore.sh removes
the packages that are already in the delivery directory. In a source-only
build this removes the binary packages of an earlier build and puts nothing
in their place.

Also, the result of the move is not examined. maker() examines only the
result of makerpm, thus a package with a failed move is not named in the
list of failed packages, and the build reports success.

Do not remove or move the binary packages when SRCONLY is set. Continue to
remove and move the source packages, because a source-only build makes
them. The three places that deliver packages get the same condition.

SRCONLY needs no other change to reach makerpm: buildcore.sh exports each
attr=val argument, thus ./buildcore.sh SRCONLY=1 gives the value to makerpm.

+++ buildcore.sh
@@ -315,6 +315,10 @@ fi
 
 setversionvars
 
+function srconly {
+    [ "$SRCONLY" = "1" -o "$SRCONLY" = "yes" ]
+}
+
 # Function for making the noarch rpms
 function maker {
     rpmname="$1"
@@ -322,9 +326,11 @@ function maker {
     if [ $? -ne 0 ]; then
         FAILEDRPMS="$FAILEDRPMS $rpmname"
     else
-        rm -f $DESTDIR/$rpmname*rpm
+        if ! srconly; then
+            rm -f $DESTDIR/$rpmname*rpm
+            mv $source/RPMS/$NOARCH/$rpmname-$VER*rpm $DESTDIR
+        fi
         rm -f $SRCDIR/$rpmname*rpm
-        mv $source/RPMS/$NOARCH/$rpmname-$VER*rpm $DESTDIR
         mv $source/SRPMS/$rpmname-$VER*rpm $SRCDIR
     fi
 }
@@ -375,9 +381,11 @@ if [ "$OSNAME" != "AIX" ]; then
             #./makerpm xCAT-genesis-scripts aarch64 "$EMBED"
             #if [ $? -ne 0 ]; then FAILEDRPMS="$FAILEDRPMS xCAT-genesis-scripts-aarch64"; fi
             if [ "$FAILEDRPMS" = "$ORIGFAILEDRPMS" ]; then    # all succeeded
-                rm -f $DESTDIR/xCAT-genesis-scripts*rpm
+                if ! srconly; then
+                    rm -f $DESTDIR/xCAT-genesis-scripts*rpm
+                    mv $source/RPMS/noarch/xCAT-genesis-scripts-*rpm $DESTDIR
+                fi
                 rm -f $SRCDIR/xCAT-genesis-scripts*rpm
-                mv $source/RPMS/noarch/xCAT-genesis-scripts-*rpm $DESTDIR
                 mv $source/SRPMS/xCAT-genesis-scripts-*rpm $SRCDIR
             fi
         fi
@@ -402,9 +410,11 @@ for rpmname in xCAT xCATsn; do
             done
         fi
         if [ "$FAILEDRPMS" = "$ORIGFAILEDRPMS" ]; then    # all succeeded
-            rm -f $DESTDIR/$rpmname-$SHORTSHORTVER*rpm
+            if ! srconly; then
+                rm -f $DESTDIR/$rpmname-$SHORTSHORTVER*rpm
+                mv $source/RPMS/*/$rpmname-$VER*rpm $DESTDIR
+            fi
             rm -f $SRCDIR/$rpmname-$SHORTSHORTVER*rpm
-            mv $source/RPMS/*/$rpmname-$VER*rpm $DESTDIR
             mv $source/SRPMS/$rpmname-$VER*rpm $SRCDIR
         fi
     fi
